+++ src/app/products/components/product-details/product-details.component.ts
@@ -1,30 +1,38 @@
 import { Component, OnInit } from '@angular/core';
-import {Observable} from 'rxjs';
-import {ProductModel} from '../../models/product.model';
-import {ActivatedRoute} from '@angular/router';
-import {ProductsService} from '../../services/products.service';
-import {CartService} from '../../../cart/services/cart.service';
+import { Observable } from 'rxjs';
+import { ProductModel } from '../../models/product.model';
+import { CartService } from '../../../cart/services/cart.service';
+import { AppState } from '../../../core/store/app.state';
+
+import { Store } from '@ngrx/store';
+
+import { getProductsByUrl } from '../../../core/store/products/products.selectors';
+import { tap } from 'rxjs/operators';
+import { loadProducts } from '../../../core/store/products/products.actions';
 
 @Component({
   selector: 'app-product-details',
   templateUrl: './product-details.component.html',
   styleUrls: ['./product-details.component.less']
 })
-export class ProductDetailsComponent {
+export class ProductDetailsComponent implements OnInit {
 
   product$: Observable<ProductModel>;
 
   constructor(
-    private route: ActivatedRoute,
-    private productsService: ProductsService,
+    private store: Store<AppState>,
     private cartService: CartService
-  ) {
-    const productID = this.route.snapshot.params.productID;
+  ) {}
 
-    if (productID) {
-      this.product$ = this.productsService.getProduct$(parseInt(productID, 10));
-    }
-
+  ngOnInit(): void {
+    this.product$ = this.store.select(getProductsByUrl)
+      .pipe(
+        tap(p => {
+          if (!p) {
+            this.store.dispatch(loadProducts());
+          }
+        })
+      );
   }
 
   onAddProductToCart(product: ProductModel): void {